Refactor SeedConsumer configuration
It is to primarily accommodate `channel` and also convert callback URL to
URL instead of string so that too many conversions are not happening.
