+++ config/test-webhook-broker.cfg
@@ -47,6 +47,8 @@ test-producer2=Test Producer 2
 
 [initial-consumers]
 test-consumer=http://imy13.us/webhook-receiver
+test-consumer6=http://imy13.us/webhook-receiver3
+test-consumer5=http://imy13.us/webhook-receiver2
 test-consumer1=./relative-url
 test-consumer2=not-a-url
 test-consumer4=http://imy13.us/webhook-receiver1
@@ -58,7 +60,20 @@ test-channel=test-channel-token
 [initial-producer-tokens]
 test-producer=test-producer-token
 
-[initial-consumer-tokens]
-test-consumer=test-consumer-token
-test-consumer1=test-consumer-token1
-test-consumer2=test-consumer-token2
+[test-consumer]
+token=test-consumer-token
+channel=test-channel
+
+[test-consumer1]
+token=test-consumer-token1
+channel=test-channel
+
+[test-consumer2]
+token=test-consumer-token2
+channel=test-channel
+
+[test-consumer4]
+channel=test-channel
+
+[test-consumer5]
+channel=non-existing-channel
